// MedicalJournal-frontend/src/app/models/MeasurementReminderEntry.ts
export class MeasurementReminderEntry{
    value1: number;
    value2: number;
    idMeasurementType: number;
    measurementTypeName: string;
    measurementValueTypeName: string;
    isDone: number;
    date: Date;
    havingMealsTime: Date;
    isLate: boolean;
    id: string;
    havingMealsType: number;
}